Reformat and clean README. Add notes about `gallery.json` to the README.
Rename "Building" section to "Deploying".
Wrap the root directory in "Deploying" in code quotes(i.e. "\`\`").
Change emphasis from underscores("_") to asterisks("*").
Change "not" in deployment warning from italics to bold.
Add a section explaining `gallery.json`, the array inside it, the elements of the array, and the properties of the elements.
Rename "Images" section to "Image Attributions."